requirements.txt
agate==1.9.1
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
backoff==2.2.1
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
choreographer==1.2.1
click==8.3.1
colorama==0.4.6
coloredlogs==14.0
comm==0.2.3
cryptography==46.0.5
daff==1.4.2
dagster==1.12.19
dagster-dbt==0.28.19
dagster-graphql==1.12.19
dagster-pipes==1.12.19
dagster-shared==1.12.19
dagster-webserver==1.12.19
db-dtypes==1.5.0
dbt-adapters==1.22.9
dbt-bigquery==1.11.1
dbt-common==1.37.3
dbt-core==1.11.7
dbt-extractor==0.6.0
dbt-protos==1.0.440
dbt-semantic-interfaces==0.9.0
debugpy==1.8.20
decorator==5.2.1
deepdiff==8.6.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
fqdn==1.5.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
google-cloud-aiplatform==1.141.0
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.2
google-cloud-core==2.5.0
google-cloud-dataproc==5.25.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.1.1
google-crc32c==1.8.0
google-genai==1.67.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
great_expectations==1.15.1
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-health-checking==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaleido==1.2.0
lark==1.3.1
leather==0.4.1
logistro==2.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mashumaro==3.14
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.1
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.4.3
oauthlib==3.3.1
orderly-set==5.5.0
orjson==3.11.7
overrides==7.7.0
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079066/work
pandas==2.3.3
pandas-gbq==0.34.0
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.6
pathlib_abc==0.5.2
pathspec==0.12.1
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
scipy==1.17.1
Send2Trash==2.1.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
snowplow-tracker==1.1.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlalchemy-bigquery==1.16.0
sqlglot==28.0.0
sqlglotrs==0.7.3
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.52.1
streamlit==1.55.0
structlog==25.5.0
tabulate==0.10.0
tenacity==9.1.4
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tomli==2.4.0
tomlkit==0.14.0
toposort==1.10
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
universal_pathlib==0.3.10
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
widgetsnbextension==4.0.15
yarl==1.23.0
zipp==3.23.0
